% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chains.R
\name{check.chains}
\alias{check.chains}
\title{check.chains}
\usage{
check.chains(X, params = c("top", "beta", "NEC"), name = "")
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from R2jags}

\item{params}{A vector of names indicating the parameters that were traced during the jags fit. For the NEC jags model this is typically 'NEC','top' and 'beta', whicha are the defaults.}

\item{name}{and optional character string indicating the label to be placed at the top of the plotting window}
}
\value{
A plot of MCMC chains and ACF diagrames for each element in params.
}
\description{
Generates a plot of MCMC chains and ACF function for a Jags model output.
}
