% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mixing.R
\name{check.mixing}
\alias{check.mixing}
\title{check.mixing}
\usage{
check.mixing(J, prob.val = 0.75)
}
\arguments{
\item{J}{a jag model fit as returned by a call to jags fit.jagsNEC}

\item{prob.val}{the probability value to pass to quantile to assess the between chain coefficient of variation against the within chain coefficient of variation. Please inspect the model code for more details.}
}
\value{
A list of within and between coefficient of variations for parameter, the outcome of the test if between is greater than within, and the ratio ratio.
}
\description{
Calculates the within and between chain mixing and tests of the between is greater than within based on the supplied quantile probability value.
}
