% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{check_inputs}
\usage{
check_inputs(data, x.var, y.var, trials.var)
}
\arguments{
\item{data}{a data.frame containing the data to use for the model}

\item{x.var}{the column heading indicating the concentration (x) variable}

\item{y.var}{the column heading indicating the response (y) variable}

\item{trials.var}{the column heading indicating the column for the number of "trials" for binomial response data.
If not supplied, the model may run but will not be the model you intended!}
}
\description{
Checks the input data for a jagsNEC model call
}
