% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modStats.R
\name{extract_modStats}
\alias{extract_modStats}
\title{extract.modStats}
\usage{
extract_modStats(mod.fits)
}
\arguments{
\item{mod.fits}{a jagsMANECfit mod.fits output list, as returned by fit.jagsMANEC}
}
\value{
A list of model statistical output derived from the input model list
}
\description{
Extracts a range of statistics from a list of jagsNECfit model fits.
}
