% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NSEC.R
\name{extract_NSEC.jagsNECfit}
\alias{extract_NSEC.jagsNECfit}
\title{extract_NSEC.jagsNSEC}
\usage{
extract_NSEC.jagsNECfit(
  X,
  sig.val = 0.01,
  precision = 10000,
  posterior = FALSE,
  hormesis.def = "control",
  x.range = NA,
  xform = NA,
  prob.vals = c(0.5, 0.025, 0.975)
)
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from fit.jagsNSEC}

\item{sig.val}{the Probability value to use as the lower quantile to test significance of the predictor posterior values
against the control, to estimate NSEC as an interpolated NOEC value from smooth ECx curves.}

\item{precision}{The number of unique x values over which to find NSEC.}

\item{posterior}{A logical value indicating if the full posterior sample of calculated NSEC values should be returned
instead of just the median and 95 credible intervals.}

\item{xform}{A function to apply to the returned estimated concentration values}

\item{prob.vals}{A vector indicating the probability values over which to return the estimated NSEC value.}
}
\value{
A vector containing the estimated NSEC value, including upper and lower 95 percent Credible Interval bounds
}
\description{
Extracts the predicted NSEC value as desired from a jagsNSEC model fit object
}
