% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NSEC.R
\name{extract_NSEC}
\alias{extract_NSEC}
\title{extract_NSEC}
\usage{
extract_NSEC(
  X,
  sig.val = 0.01,
  precision = 10000,
  posterior = FALSE,
  type = "absolute",
  hormesis.def = "control",
  xform = NA,
  x.range = NA,
  prob.vals = c(0.5, 0.025, 0.975),
  link = "identity"
)
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from fit.jagsNSEC}

\item{sig.val}{the Probability value to use as the lower quantile to test significance of the predictor posterior values
against the control, to estimate NSEC as an interpolated NOEC value from smooth ECx curves.}

\item{precision}{The number of unique x values over which to find NSEC - large values will make the NSEC estimate more 
precise.}

\item{posterior}{A logical value indicating if the full posterior sample of calculated NSEC values should be returned 
instead of just the median and 95 credible intervals.}

\item{xform}{A function to apply to the returned estimated concentration values}

\item{x.range}{A range of x values over which to consider extracting NSEC}

\item{prob.vals}{A vector indicating the probability values over which to return the estimated NSEC value. Defaults to 0.5 (median) and 0.025 and 0.975 (95 percent credible intervals).}
}
\value{
A vector containing the estimated NSEC value, including upper and lower 95 percent Credible Interval bounds
}
\description{
Extracts the predicted NSEC value as desired from a jagsNSEC or a jagsMANSEC model fit.
}
