% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jagsMod.R
\name{fit.jagsMod}
\alias{fit.jagsMod}
\title{fit.jagsMod}
\usage{
fit.jagsMod(
  mod.dat,
  init.fun,
  params,
  burnin,
  n.iter,
  init.value.warning,
  n.tries
)
}
\arguments{
\item{mod.dat}{the model data to use for the model fit}

\item{init.fun}{the init function to use for the model fit}

\item{params}{a vector of names indicating the parameters to trace during the jags fit.}

\item{burnin}{the number of iterations to use as burnin.}

\item{n.iter}{the number of iterations to run following burnin for the initial jags fit. Defaults to 500 + burnin,}

\item{init.value.warning}{indicates if a warning should be given if the init.fun generated fails.}

\item{n.tries}{the number of tries to attempt to fit the model and attain good chain mixing. See details below.}
}
\value{
The fitted jags model output
}
\description{
Fits a Bayesian model using jags and R2jags
}
