% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_jagsNEC.R
\name{fit.jagsNEC}
\alias{fit.jagsNEC}
\title{fit.jagsNEC}
\usage{
fit.jagsNEC(data, x.var, y.var, trials.var = NA, x.type = "gamma",
  y.type, params = c("top", "beta", "NEC"), burnin = 1000,
  n.iter = burnin * 2)
}
\arguments{
\item{data}{a data.frame containing the data to use for the model}

\item{x.var}{the column heading indicating the concentration (x) variable}

\item{y.var}{the column heading indicating the response (y) variable}

\item{trials.var}{the column heading indicating the column for the number of "trials" for binomial response data}

\item{x.type}{the statistical distribution to use for the x (concentration) data. The default is "gamma" as most concentration data are gamma (ie continuous on the scale of >0 to inf). As some experiments will use zero concentration, and there is no distribution on the continuous scale from 0 to in (ie tweedie) available in jags, a small offset is added (1/10^5 of the next lowest value) to zero values of concentration}

\item{y.type}{the statistical distribution to use for the y (response) data. This may currently be one of  'binomial', 'poisson', or 'gamma'. Others can be added as required, please contact the package maintainer.}

\item{params}{A vector of names indicating the parameters that were traced during the jags fit. For the NEC jags model this is typically 'NEC','top' and 'beta', whicha are the defaults.}

\item{burnin}{the number of iterations to use as burning.}

\item{n.iter}{the number of interations to run overall. Defaults to 2 times burnin.}
}
\value{
The $BUGSoutput element of fitted jags model.
}
\description{
Fits an NEC model as per Fox 2010, using jags and R2jags
}
