% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsNEC_input.R
\name{jagsNEC_input}
\alias{jagsNEC_input}
\title{jagsNEC_input}
\usage{
jagsNEC_input(
  data,
  x.var,
  y.var,
  trials.var,
  x.type,
  y.type,
  params,
  over.disp,
  model
)
}
\arguments{
\item{data}{a data.frame containing the data to use for the model}

\item{x.var}{the column heading indicating the concentration (x) variable}

\item{y.var}{the column heading indicating the response (y) variable}

\item{trials.var}{the column heading indicating the column for the number of "trials" for binomial response data.
If not supplied, the model may run but will not be the model you intended!}

\item{x.type}{the statistical distribution to use for the x (concentration) data. This will be guess based on the
characteristic of the input data if not supplied.}

\item{y.type}{the statistical distribution to use for the y (response) data. This may currently be one of  'binomial',
'poisson',' 'gaussian', 'beta', 'negbin' or 'gamma'. Others can be added as required, please contact the package maintainer.
If not supplied, the appropriate distribution will be guessed based on the distribution of the input data.}

\item{params}{a vector of names indicating the parameters to trace during the jags fit.}

\item{model}{the type of model to be fit. Currently takes values of "NEC3param",
"NEC4param", "NECsigmoidal", "NECHormesis", "ECx4param", "ECxExp", "ECxLinear", "ECxsigmoidal", "ECxWeibull1", or "ECxWeibull2".}
}
\value{
Modified elements of the jagsNEC input data.
}
\description{
Checks the model inputs for a jagsNEC model fit
}
\details{
This is a wrapper function to test input data criteria and write the jags model file for use in a jagsNEC model fit
}
