% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_jagsMANEC.R
\name{modify.jagsMANEC}
\alias{modify.jagsMANEC}
\title{modify.jagsMANEC}
\usage{
\method{modify}{jagsMANEC}(jagsMANECfit, model.set = NA)
}
\arguments{
\item{jagsMANECfit}{a jagsMANECfit output list, as returned by fit.jagsMANEC}

\item{model.set}{A character vector containing the of names of model types to be included in the modified fit.}
}
\value{
All successully fitted jagsMANECfit model fit.
}
\description{
Modifys an existing jagsMANECfit, for example, but adding or removing fitted models.
}
