% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jagsNECfit.R
\name{plot.jagsMANECfit}
\alias{plot.jagsMANECfit}
\title{plot.jagsMANEC}
\usage{
\method{plot}{jagsMANECfit}(
  X,
  CI = TRUE,
  posterior.median = TRUE,
  median.model = FALSE,
  add.NEC = TRUE,
  add.EC10 = FALSE,
  legend.loc = "topright",
  xform = NA,
  lxform = NA,
  jitter.x = FALSE,
  jitter.y = FALSE,
  ylab = "response",
  xlab = "concentration",
  xlim = NA,
  ylim = NA,
  xticks = NA,
  xtick.rounding = 5,
  all_models = FALSE,
  ...
)
}
\arguments{
\item{X}{the jagsNEC model fit as returned by fit.jagsNEC.}

\item{CI}{a logical value indicating if confidence intervals on the model fit should be plotted,
calculated as the upper and lower bounds of the individual predicted values from all posterior samples}

\item{posterior.median}{a logical value indicating if the posterior median of the model fit should be plotted,
calculated as the median of the individual predicted values from all posterior samples}

\item{median.model}{a logical value indicating if the fitted model calculated from the median estimates of the NEC,
top and beta parameters should be plotted. This is the fitted model as shown in Fox 2010.}

\item{add.NEC}{a logical value indicating if the estimated NEC value and 95\% credible intervals should be added to the plot.}

\item{add.EC10}{a logical value indicating if an estimated EC10 value and 95\% credible intervals should be added to the plot.}

\item{legend.loc}{a vector indicating the location of the NEC or EC10 legend, as per a call to legend.}

\item{xform}{a function to be applied as a transformation of the x data.}

\item{lxform}{a function to be applied as a transformation only to axis labels and the annoted NEC/EC10 values.}

\item{jitter.x}{a logical value indicating if the x data points on the plot should be jittered.}

\item{jitter.y}{a logical value indicating if the y data points on the plot should be jittered.}

\item{ylab}{a character vector to use for the y-axis label}

\item{xlab}{a character vector to use for the x-axis label}

\item{xlim}{a numeric vector of length two to use for the lower and uper limits of the x-axis range}

\item{xticks}{a numeric vector indicating where to place the tick marks of the x-axis}

\item{xtick.rounding}{the rounding to be applied for manually specifying xticks, when data are back transformed via lxform.}

\item{all_models}{a logical, indicating if all models should be plotted individually, or if the model averaged plot should be returned.}
}
\value{
a plot of the fitted model
}
\description{
Generates a plot of a fitted jagsMANEC model, as returned by fit.jagsMANEC.
}
