% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jagsNECfit.R
\name{plot.jagsNECfit}
\alias{plot.jagsNECfit}
\title{plot.jagsNEC}
\usage{
\method{plot}{jagsNECfit}(X, CI = TRUE, posterior.median = TRUE,
  median.model = FALSE, add.NEC = TRUE, xform = NA,
  jitter.x = FALSE, jitter.y = FALSE, y.lab = "response",
  x.lab = "concentration", log.x = "", ...)
}
\arguments{
\item{X}{the jags NEC model fit as returned by fit.jagsNEC.}

\item{CI}{a logical value indicating if confidence intervals on the model fit should be plotted, calculated as the upper and lower bounds of the individual predicted values from all posterior samples}

\item{posterior.median}{a logical value indicating if the posterior median of the model fit should be plotted, calculated as the median of the individual predicted values from all posterior samples}

\item{median.model}{a logical value indicating if the fitted model calculated from the median estimates of the NEC, top and beta parameters should be plotted. This is the fitted model as shown in Fox 2010.}

\item{add.NEC}{a logical value indicating if the estimated NEC values and 95\% credible intervals should be added to the plot.}

\item{xform}{a function to be applied as a transformation of the x data.}

\item{y.lab}{A character string containing a label for the y-axis label}

\item{x.lab}{A character string containing a label for the x-axis}

\item{log.x}{A character string indicating if the x-axis should be plotted on a log scale. Leave as the default for a plot without a log x-axis, or "x" will provide a plot with the x-axis on the log scale.}

\item{x.jitter}{a logical value indicating if the x data points on the plot should be jittered.}

\item{y.jitter}{a logical value indicating if the y data points on the plot should be jittered.}
}
\value{
a plot of the fitted model
}
\description{
Generates a plot of a fitted jags NEC model, as returned by fit.jagsNEC.
}
