% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jagsNEC.R
\name{plot_jagsNEC}
\alias{plot_jagsNEC}
\title{plot.jagsNEC}
\usage{
plot_jagsNEC(X, CI = TRUE, posterior.median = TRUE,
  median.model = FALSE, add.NEC = TRUE)
}
\arguments{
\item{X}{the jags NEC model fit as returned by fit.jagsNEC.}

\item{CI}{a logical value indicating if confidence intervals on the model fit should be plotted, calculated as the upper and lower bounds of the individual predicted values from all posterior samples}

\item{posterior.median}{a logical value indicating if the posterior median of the model fit should be plotted, calculated as the median of the individual predicted values from all posterior samples}

\item{median.model}{a logical value indicating of the fitted model caluclated from the median estimates of the NEC, top and beta parameters should be plotted. This is the fitted model as shown in Fox 2010.}

\item{add.NEC}{a logocal value indicating if the estimated NEC values and 95\% credible intervals should be added to the plot.}
}
\value{
a plot of the fitted model
}
\description{
Generates a plot of a fitted jags NEC model, as returned by fit.jagsNEC.
}
