% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_fitted_vals.R
\name{predict_Linearmod}
\alias{predict_Linearmod}
\title{predict_Linearmod}
\usage{
predict_Linearmod(x.vec, top, beta, link)
}
\arguments{
\item{x.vec}{the x vector over which to calculate}

\item{top}{the upper plateau}

\item{beta}{the linear decay rate}

\item{link}{the link function used}
}
\value{
A list containing x and fitted y, with up and lw values
}
\description{
Calculates predicted y (response) values for a supplied vector of x (concentration) values
}
