% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wi.R
\name{wi}
\alias{wi}
\title{wi}
\usage{
wi(vec)
}
\arguments{
\item{vec}{A vector of AIC, AICc or BIC values}
}
\value{
A vector of weights
}
\description{
Calculates model weights based on AIC, AICc, or BIC, or DIC
}
