% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_jags_ECxLinearMod.R
\name{write.jags.ECxLinear.mod}
\alias{write.jags.ECxLinear.mod}
\title{write.jags.ECxLinear.mod}
\usage{
write.jags.ECxLinear.mod(x = "gamma", y, mod.dat)
}
\arguments{
\item{x}{the statistical distribution to use for the x (concentration) data. This may currently be one of  'beta', 'gaussian', or 'gamma'. Others can be added as required, please contact the package maintainer.}

\item{y}{the statistical distribution to use for the y (response) data. This may currently be one of  'binomial', 'beta', 'poisson', 'gaussian', or 'gamma'. Others can be added as required, please contact the package maintainer.}

\item{mod.dat}{the model data to use}
}
\value{
an init function to pass to jags
}
\description{
Writes a ECxLinear file and generates a function for initial values to pass to jags
}
