% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_jags_NEC3paramMod.R
\name{write.jags.NEC3param.mod}
\alias{write.jags.NEC3param.mod}
\title{write.jags.NEC3param.mod}
\usage{
write.jags.NEC3param.mod(x = "gamma", y, mod.dat)
}
\arguments{
\item{x}{the statistical distribution to use for the x (concentration) data. This may currently be one of  'beta', 'gaussian', or 'gamma'. Others can be added as required, please contact the package maintainer.}

\item{y}{the statistical distribution to use for the y (response) data. This may currently be one of  'binomial', 'beta', 'poisson', 'gaussian', or 'gamma'. Others can be added as required, please contact the package maintainer.}
}
\value{
an init function to pass to jags
}
\description{
Writes an NEC model file for a three parameter model (top, beta and NEC) and generates a function for initial values to pass to jags
}
