% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_jags_model.R
\name{write.jags.NECmod}
\alias{write.jags.NECmod}
\title{write.jags.NECmod}
\usage{
write.jags.NECmod(x = "gamma", y)
}
\arguments{
\item{x}{the statistical distribution to use for the x (concentration) data. This may currently be one of  'beta', 'gaussian', or 'gamma'. Others can be added as required, please contact the package maintainer.}

\item{y}{the statistical distribution to use for the y (response) data. This may currently be one of  'binomial', 'poisson', or 'gamma'. Others can be added as required, please contact the package maintainer.}
}
\value{
an init function to pass to jags
}
\description{
Writes and NEC model file and generates a function for initial values to pass to jags
}
