% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_jagsModFile.R
\name{Write.jagsModFile}
\alias{Write.jagsModFile}
\title{Write.jagsModFile}
\usage{
Write.jagsModFile(x.type, y.type, mod.dat, params, model)
}
\arguments{
\item{x.type}{The type of predictor data to write the model for}

\item{y.type}{The type of response data to write the model for}

\item{mod.dat}{A list containing the jags model data}

\item{params}{A character vector containing the parameters of the model}

\item{model}{The type of model to fit}
}
\value{
In init.fun and updated params vector
}
\description{
Writes a jags model files and generates the required init.fun and param vector.
}
