% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_fitted_vals.R
\name{xform_link}
\alias{xform_link}
\title{xform_link}
\usage{
xform_link(pred.vec, use.link)
}
\arguments{
\item{x.vec}{the x vector over which to calculate}

\item{link}{The link function used}
}
\value{
A list containing x and fitted y, with up and lw values
}
\description{
Calculates predicted y (response) values for a supplied vector of x (concentration) values
}
