% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ECx.R
\name{extract_ECx}
\alias{extract_ECx}
\title{extract_ECx}
\usage{
extract_ECx(X, ECx.val = 10)
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from fit.jagsNEC}

\item{ECx.val}{the desired percentage effect value. This must be a value between 1 and 99, defaults to 10.}
}
\value{
A vector containing the estimated ECx value, including upper and lower 95% Credible Interval bounds
}
\description{
Extracts the predicted ECx value as desired from a jagsNEC model fit obeject
}
