% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ECx.R
\name{extract_ECx}
\alias{extract_ECx}
\title{extract_ECx}
\usage{
extract_ECx(X, ECx.val = 10, precision = 10000, posterior = FALSE,
  type = "absolute", xform = NA)
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from fit.jagsNEC}

\item{ECx.val}{the desired percentage effect value. This must be a value between 1 and 99 (for type = "relative" and "absolute"), defaults to 10.}

\item{precision}{The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.}

\item{posterior}{A logical value indicating if the full posterior sample of calculated ECx values should be returned instead of just the median and 95 credible intervals.}

\item{type}{a character vector indicating if relative or absolute values for the ECx should be calculated. Takes values 
of "relative",  "absolute" (the default) or "direct". For the default model type fit by jagsNEC, "relative" is 
calculated as the percentage decrease from the maximum value of the response (top) to the minimum predicted value 
of the response, "absolute" is calculated as the the percentage decrease from the maximum value of the response (top) 
to 0 (or bot for a 4 parameter model fit),  and "direct" provides a direct estimate of the x value for a given y.}

\item{xform}{A function to apply to the returned estimated concentration values}
}
\value{
A vector containing the estimated ECx value, including upper and lower 95 percent Credible Interval bounds
}
\description{
Extracts the predicted ECx value as desired from a jagsNEC model fit obeject
}
\details{
Please note that the estimated ECx value is based on the equivalent percentage decrease from the range of the highest to the lowest estimate value across the range of the observed concentration (x) values. If the concentration response relationship is such that the full range of observed responses is not captured (ie a complete decline response at the highest level of exposure), the estimated ECx values may be lower than if the full concentration-response curve were available. Note this is therefore a conservative value.
}
