% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jagsMod.R
\name{fit.jagsMod}
\alias{fit.jagsMod}
\title{fit.jagsMod}
\usage{
fit.jagsMod(
  mod.dat,
  init.fun,
  params,
  burnin,
  n.iter,
  init.value.warning,
  n.tries
)
}
\arguments{
\item{mod.dat}{The jags model data to use}

\item{init.fun}{A jags init.fun with which to generate starting values}

\item{params}{A vector of names indicating the parameters that to trace during the jags fit}

\item{burnin}{the number of interations to run overall. This should be large. Defaults to 10000.}

\item{n.iter}{the number of interations to run following burnin for the initial jags fit. Defaults to 500 + burnin,}

\item{init.value.warning}{Indicates if a warning should be given if the init.fun generated fails.}

\item{n.tries}{The number of tries to attempt to fit the model and attain good chain mixing. See details below.}
}
\value{
The fitted jags model output
}
\description{
Fits an NEC model as per Fox 2010, using jags and R2jags
}
