% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jagsNECfit.R
\name{plot.jagsNECfit}
\alias{plot.jagsNECfit}
\title{plot.jagsNEC}
\usage{
\method{plot}{jagsNECfit}(X, CI = TRUE, posterior.median = TRUE,
  median.model = FALSE, add.NEC = TRUE, NEC.loc = "topright",
  xform = NA, lxform = NA, jitter.x = FALSE, jitter.y = FALSE,
  ylab = "response", xlab = "concentration", ...)
}
\arguments{
\item{X}{the jags NEC model fit as returned by fit.jagsNEC.}

\item{CI}{a logical value indicating if confidence intervals on the model fit should be plotted, calculated as the upper and lower bounds of the individual predicted values from all posterior samples}

\item{posterior.median}{a logical value indicating if the posterior median of the model fit should be plotted, calculated as the median of the individual predicted values from all posterior samples}

\item{median.model}{a logical value indicating if the fitted model calculated from the median estimates of the NEC, top and beta parameters should be plotted. This is the fitted model as shown in Fox 2010.}

\item{add.NEC}{a logical value indicating if the estimated NEC values and 95\% credible intervals should be added to the plot.}

\item{NEC.loc}{a vector indicating the location of the NEC, as per a call to legend.}

\item{xform}{a function to be applied as a transformation of the x data.}

\item{lxform}{a function to be applied as a transformation only to axis labels and the annoted NEC value.}

\item{x.jitter}{a logical value indicating if the x data points on the plot should be jittered.}

\item{y.jitter}{a logical value indicating if the y data points on the plot should be jittered.}
}
\value{
a plot of the fitted model
}
\description{
Generates a plot of a fitted jags NEC model, as returned by fit.jagsNEC.
}
