% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_fitted_vals.R
\name{predict_NECbugsmod}
\alias{predict_NECbugsmod}
\title{predict_NECbugsmod}
\usage{
predict_NECbugsmod(X, precision = 100)
}
\arguments{
\item{X}{the jags NEC model fit (as returned by fit.jagsNEC)}

\item{precision}{the number of x values over which to predict values}
}
\value{
a list containing x (the x.seq), y (the median predicted y values at each value of x), up.vals (the upper 97.5% percentile of the predicted y values at each value of x), and lw (the lower 2.5% percentile of the predicted y values at each value of x)
}
\description{
Calculates predicted y (response) values for a supplied vector of x (concentration) values for a jags fitted NEC model fit (as returned by fit.jagsNEC)
}
