% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_fitted_vals.R
\name{predict_NECmod}
\alias{predict_NECmod}
\title{predict_NECmod}
\usage{
predict_NECmod(
  x.vec,
  NEC = min(x.vec),
  top,
  beta,
  alpha = 0,
  d = 1,
  bot = 0,
  slope = 0
)
}
\arguments{
\item{x.vec}{the x vector over which to calculate}

\item{NEC}{the No-Effect-Concentration}

\item{top}{the estimated value of y (response) over the range of x for which no effect occurs}

\item{beta}{the exponential decay rate}

\item{alpha}{the offset of a gaussian y response variable}

\item{bot}{the lower plateau}
}
\value{
A list containing x and fitted y, with up and lw values
}
\description{
Calculates predicted y (response) values for a supplied vector of x (concentration) values
}
